requirements.txt
streamlit
requests
beautifulsoup4
google-generativeai
python-dotenv 